[Enhancement #206] Refactor Jena list handlers in preparation for adding support for data property referenced list values.
