+++ ontodriver-jena/src/main/java/cz/cvut/kbss/ontodriver/jena/list/JenaLists.java
@@ -75,14 +75,14 @@ public class JenaLists implements Lists {
     }
 
     @Override
-    public void persistReferencedList(ReferencedListValueDescriptor descriptor) throws OntoDriverException {
+    public <V> void persistReferencedList(ReferencedListValueDescriptor<V> descriptor) throws OntoDriverException {
         executeBeforeCallbackAndVerifyArgument(descriptor);
         adapter.referencedListHandler().persistList(descriptor);
         afterChangeCallback.execute();
     }
 
     @Override
-    public void updateReferencedList(ReferencedListValueDescriptor descriptor) throws OntoDriverException {
+    public <V> void updateReferencedList(ReferencedListValueDescriptor<V> descriptor) throws OntoDriverException {
         executeBeforeCallbackAndVerifyArgument(descriptor);
         adapter.referencedListHandler().updateList(descriptor);
         afterChangeCallback.execute();
