+++ ontodriver-jena/src/main/java/cz/cvut/kbss/ontodriver/jena/list/SimpleListHandler.java
@@ -20,6 +20,7 @@ package cz.cvut.kbss.ontodriver.jena.list;
 import cz.cvut.kbss.ontodriver.descriptor.SimpleListDescriptor;
 import cz.cvut.kbss.ontodriver.descriptor.SimpleListValueDescriptor;
 import cz.cvut.kbss.ontodriver.jena.connector.StorageConnector;
+import cz.cvut.kbss.ontodriver.model.Axiom;
 import cz.cvut.kbss.ontodriver.model.NamedResource;
 import org.apache.jena.rdf.model.Property;
 import org.apache.jena.rdf.model.Resource;
@@ -32,23 +33,32 @@ import java.util.List;
 
 import static org.apache.jena.rdf.model.ResourceFactory.createResource;
 
-class SimpleListHandler extends ListHandler<SimpleListDescriptor, SimpleListValueDescriptor> {
+public class SimpleListHandler {
 
-    SimpleListHandler(StorageConnector connector) {
-        super(connector);
+    private final StorageConnector connector;
+
+    public SimpleListHandler(StorageConnector connector) {
+        this.connector = connector;
     }
 
-    @Override
-    SimpleListIterator iterator(SimpleListDescriptor descriptor) {
-        return new SimpleListIterator(descriptor, connector);
+    List<Axiom<NamedResource>> loadList(SimpleListDescriptor descriptor) {
+        final List<Axiom<NamedResource>> result = new ArrayList<>();
+        final AbstractListIterator it = new SimpleListIterator(descriptor, connector);
+        while (it.hasNext()) {
+            result.add(it.nextAxiom());
+        }
+        return result;
     }
 
-    @Override
-    AbstractListIterator iterator(SimpleListValueDescriptor descriptor) {
-        return iterator((SimpleListDescriptor) descriptor);
+    void persistList(SimpleListValueDescriptor descriptor) {
+        final List<NamedResource> values = descriptor.getValues();
+        if (values.isEmpty()) {
+            return;
+        }
+        Resource owner = createResource(descriptor.getListOwner().getIdentifier().toString());
+        appendNewNodes(descriptor, 0, owner);
     }
 
-    @Override
     void appendNewNodes(SimpleListValueDescriptor descriptor, int index, Resource lastNode) {
         final Property hasList = ResourceFactory
                 .createProperty(descriptor.getListProperty().getIdentifier().toString());
@@ -66,5 +76,29 @@ class SimpleListHandler extends ListHandler<SimpleListDescriptor, SimpleListValu
         final Resource node = createResource(value.getIdentifier().toString());
         statements.add(ResourceFactory.createStatement(previous, property, node));
         return node;
+    }
+
+    void updateList(SimpleListValueDescriptor descriptor) {
+        final AbstractListIterator it = new SimpleListIterator(descriptor, connector);
+        int i = 0;
+        while (it.hasNext() && i < descriptor.getValues().size()) {
+            final NamedResource update = descriptor.getValues().get(i);
+            final NamedResource existing = it.nextValue();
+            if (!existing.equals(update)) {
+                it.replace(createResource(update.getIdentifier().toString()));
+            }
+            i++;
+        }
+        removeObsoleteNodes(it);
+        if (i < descriptor.getValues().size()) {
+            appendNewNodes(descriptor, i, it.getCurrentNode());
+        }
+    }
+
+    private static void removeObsoleteNodes(AbstractListIterator it) {
+        while (it.hasNext()) {
+            it.nextValue();
+            it.removeWithoutReconnect();
+        }
     }
 }
