--- ontodriver-jena/src/test/java/cz/cvut/kbss/ontodriver/jena/list/ListHandlerTestBase.java
@@ -1,99 +1,0 @@
-/*
- * JOPA
- * Copyright (C) 2023 Czech Technical University in Prague
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3.0 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library.
- */
-package cz.cvut.kbss.ontodriver.jena.list;
-
-import cz.cvut.kbss.ontodriver.descriptor.ListDescriptor;
-import cz.cvut.kbss.ontodriver.descriptor.ListValueDescriptor;
-import cz.cvut.kbss.ontodriver.jena.connector.StorageConnector;
-import cz.cvut.kbss.ontodriver.jena.environment.Generator;
-import cz.cvut.kbss.ontodriver.model.Assertion;
-import cz.cvut.kbss.ontodriver.model.Axiom;
-import cz.cvut.kbss.ontodriver.model.NamedResource;
-import org.apache.jena.rdf.model.Property;
-import org.apache.jena.rdf.model.Resource;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-
-import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static org.apache.jena.rdf.model.ResourceFactory.createProperty;
-import static org.apache.jena.rdf.model.ResourceFactory.createResource;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-
-abstract class ListHandlerTestBase<D extends ListDescriptor, V extends ListValueDescriptor> {
-
-    static final NamedResource OWNER = NamedResource.create(Generator.generateUri());
-    static final Assertion HAS_LIST = Assertion.createObjectPropertyAssertion(Generator.generateUri(), false);
-    static final Assertion HAS_NEXT = Assertion.createObjectPropertyAssertion(Generator.generateUri(), false);
-    static final Resource OWNER_RESOURCE = createResource(OWNER.getIdentifier().toString());
-    static final Property HAS_LIST_PROPERTY = createProperty(HAS_LIST.getIdentifier().toString());
-    static final Property HAS_NEXT_PROPERTY = createProperty(HAS_NEXT.getIdentifier().toString());
-
-    @Mock
-    StorageConnector connectorMock;
-
-    ListTestUtil listUtil;
-
-    ListHandler<D, V> handler;
-
-    public void setUp() {
-        this.listUtil = new ListTestUtil(OWNER_RESOURCE, HAS_LIST_PROPERTY, HAS_NEXT_PROPERTY, connectorMock);
-    }
-
-    @Test
-    public void loadListRetrievesAllListElements() {
-        final List<URI> expected = generateList(null);
-        final List<Axiom<NamedResource>> result = handler.loadList(listDescriptor());
-        assertNotNull(result);
-        final List<URI> actual = result.stream().map(ax -> ax.getValue().getValue().getIdentifier())
-                                       .collect(Collectors.toList());
-        assertEquals(expected, actual);
-    }
-
-    abstract List<URI> generateList(String context);
-
-    abstract D listDescriptor();
-
-    @Test
-    public void loadListFromContextRetrievesAllListElements() {
-        final URI context = Generator.generateUri();
-        final List<URI> expected = generateList(context.toString());
-        final D descriptor = listDescriptor();
-        descriptor.setContext(context);
-        final List<Axiom<NamedResource>> result = handler.loadList(descriptor);
-        final List<URI> actual = result.stream().map(ax -> ax.getValue().getValue().getIdentifier())
-                                       .collect(Collectors.toList());
-        assertEquals(expected, actual);
-    }
-
-    abstract V listValueDescriptor();
-
-    @Test
-    public void persistListDoesNothingForEmptyDescriptor() {
-        final V descriptor = listValueDescriptor();
-        handler.persistList(descriptor);
-        verify(connectorMock, never()).add(anyList(), anyString());
-    }
-}
