+++ ontodriver-jena/src/test/java/cz/cvut/kbss/ontodriver/jena/list/ReferencedListHandlerTest.java
@@ -22,6 +22,7 @@ import cz.cvut.kbss.ontodriver.descriptor.ReferencedListDescriptorImpl;
 import cz.cvut.kbss.ontodriver.descriptor.ReferencedListValueDescriptor;
 import cz.cvut.kbss.ontodriver.jena.environment.Generator;
 import cz.cvut.kbss.ontodriver.model.Assertion;
+import cz.cvut.kbss.ontodriver.model.Axiom;
 import cz.cvut.kbss.ontodriver.model.NamedResource;
 import org.apache.jena.rdf.model.Property;
 import org.apache.jena.rdf.model.Resource;
@@ -38,22 +39,30 @@ import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
-import static org.apache.jena.rdf.model.ResourceFactory.*;
+import static org.apache.jena.rdf.model.ResourceFactory.createProperty;
+import static org.apache.jena.rdf.model.ResourceFactory.createResource;
+import static org.apache.jena.rdf.model.ResourceFactory.createStatement;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
-public class ReferencedListHandlerTest
-        extends ListHandlerTestBase<ReferencedListDescriptor, ReferencedListValueDescriptor> {
+public class ReferencedListHandlerTest extends ListHandlerTestHelper {
 
     private static final Assertion HAS_CONTENT = Assertion
             .createObjectPropertyAssertion(Generator.generateUri(), false);
     private static final Property HAS_CONTENT_PROPERTY = createProperty(HAS_CONTENT.getIdentifier().toString());
+
+    private ReferencedListHandler handler;
 
     @BeforeEach
     public void setUp() {
@@ -62,7 +71,6 @@ public class ReferencedListHandlerTest
         listUtil.setHasContent(HAS_CONTENT_PROPERTY);
     }
 
-    @Override
     List<URI> generateList(String context) {
         if (context != null) {
             return listUtil.generateReferencedList(context);
@@ -71,20 +79,47 @@ public class ReferencedListHandlerTest
         }
     }
 
-    @Override
     ReferencedListDescriptor listDescriptor() {
         return new ReferencedListDescriptorImpl(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
     }
 
-    @Override
-    ReferencedListValueDescriptor listValueDescriptor() {
-        return new ReferencedListValueDescriptor(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+    ReferencedListValueDescriptor<NamedResource> listValueDescriptor() {
+        return new ReferencedListValueDescriptor<>(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+    }
+
+    @Test
+    public void loadListRetrievesAllListElements() {
+        final List<URI> expected = generateList(null);
+        final List<Axiom<NamedResource>> result = handler.loadList(listDescriptor());
+        assertNotNull(result);
+        final List<URI> actual = result.stream().map(ax -> ax.getValue().getValue().getIdentifier())
+                                       .collect(Collectors.toList());
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void loadListFromContextRetrievesAllListElements() {
+        final URI context = Generator.generateUri();
+        final List<URI> expected = generateList(context.toString());
+        final ReferencedListDescriptor descriptor = listDescriptor();
+        descriptor.setContext(context);
+        final List<Axiom<NamedResource>> result = handler.loadList(descriptor);
+        final List<URI> actual = result.stream().map(ax -> ax.getValue().getValue().getIdentifier())
+                                       .collect(Collectors.toList());
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void persistListDoesNothingForEmptyDescriptor() {
+        final ReferencedListValueDescriptor<NamedResource> descriptor = listValueDescriptor();
+        handler.persistList(descriptor);
+        verify(connectorMock, never()).add(anyList(), anyString());
     }
 
     @Test
     public void persistListInsertsStatementsCorrespondingToList() {
         final List<URI> list = listUtil.generateList();
-        final ReferencedListValueDescriptor descriptor = listValueDescriptor();
+        final ReferencedListValueDescriptor<NamedResource> descriptor = listValueDescriptor();
         list.forEach(item -> descriptor.addValue(NamedResource.create(item)));
         handler.persistList(descriptor);
         final ArgumentCaptor<List> captor = ArgumentCaptor.forClass(List.class);
@@ -103,7 +138,7 @@ public class ReferencedListHandlerTest
     @Test
     public void persistListWithContextInsertsStatementsCorrespondingToListIntoContext() {
         final List<URI> list = listUtil.generateList();
-        final ReferencedListValueDescriptor descriptor = listValueDescriptor();
+        final ReferencedListValueDescriptor<NamedResource> descriptor = listValueDescriptor();
         list.forEach(item -> descriptor.addValue(NamedResource.create(item)));
         final URI context = Generator.generateUri();
         descriptor.setContext(context);
@@ -118,7 +153,7 @@ public class ReferencedListHandlerTest
     public void updateListAddsNodeWithContentToEnd() {
         final List<URI> list = generateList(null);
         final URI newItem = Generator.generateUri();
-        final ReferencedListValueDescriptor descriptor = listValueDescriptor();
+        final ReferencedListValueDescriptor<NamedResource> descriptor = listValueDescriptor();
         list.forEach(item -> descriptor.addValue(NamedResource.create(item)));
         descriptor.addValue(NamedResource.create(newItem));
         handler.updateList(descriptor);
@@ -136,8 +171,8 @@ public class ReferencedListHandlerTest
     @Test
     public void updateListRemovesRemainingExistingNodes() {
         final List<URI> list = generateList(null);
-        final ReferencedListValueDescriptor descriptor =
-                new ReferencedListValueDescriptor(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+        final ReferencedListValueDescriptor<NamedResource> descriptor =
+                new ReferencedListValueDescriptor<>(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
         final List<URI> update = list.subList(0, list.size() / 2);
         update.forEach(item -> descriptor.addValue(NamedResource.create(item)));
         handler.updateList(descriptor);
@@ -152,8 +187,8 @@ public class ReferencedListHandlerTest
     @Test
     public void updateReplacesModifiedNodeContent() {
         final List<URI> list = generateList(null);
-        final ReferencedListValueDescriptor descriptor =
-                new ReferencedListValueDescriptor(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+        final ReferencedListValueDescriptor<NamedResource> descriptor =
+                new ReferencedListValueDescriptor<>(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
         final List<URI> update = new ArrayList<>(list);
         final URI replace = Generator.generateUri();
         final int index = Generator.randomInt(list.size());
@@ -170,8 +205,8 @@ public class ReferencedListHandlerTest
     @Test
     public void updateClearsOriginalListWhenUpdateIsEmpty() {
         final List<URI> list = generateList(null);
-        final ReferencedListValueDescriptor descriptor =
-                new ReferencedListValueDescriptor(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+        final ReferencedListValueDescriptor<NamedResource> descriptor =
+                new ReferencedListValueDescriptor<>(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
         handler.updateList(descriptor);
 
         for (int i = 0; i < list.size(); i++) {
@@ -189,8 +224,8 @@ public class ReferencedListHandlerTest
 
     @Test
     public void updatePersistsListWhenOriginalWasEmpty() {
-        final ReferencedListValueDescriptor descriptor =
-                new ReferencedListValueDescriptor(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+        final ReferencedListValueDescriptor<NamedResource> descriptor =
+                new ReferencedListValueDescriptor<>(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
         final List<URI> list = listUtil.generateList();
         list.forEach(u -> descriptor.addValue(NamedResource.create(u)));
         handler.updateList(descriptor);
@@ -205,8 +240,8 @@ public class ReferencedListHandlerTest
     public void updateListWorksInContext() {
         final URI context = Generator.generateUri();
         final List<URI> list = generateList(context.toString());
-        final ReferencedListValueDescriptor descriptor =
-                new ReferencedListValueDescriptor(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
+        final ReferencedListValueDescriptor<NamedResource> descriptor =
+                new ReferencedListValueDescriptor<>(OWNER, HAS_LIST, HAS_NEXT, HAS_CONTENT);
         descriptor.setContext(context);
         final URI firstReplaced = Generator.generateUri();
         descriptor.addValue(NamedResource.create(firstReplaced));
