+++ ontodriver-rdf4j/src/test/java/cz/cvut/kbss/ontodriver/rdf4j/ListHandlerTestHelper.java
@@ -1,18 +1,14 @@
 package cz.cvut.kbss.ontodriver.rdf4j;
 
-import cz.cvut.kbss.ontodriver.model.Axiom;
 import cz.cvut.kbss.ontodriver.model.NamedResource;
 import cz.cvut.kbss.ontodriver.rdf4j.environment.Generator;
 import cz.cvut.kbss.ontodriver.rdf4j.environment.Vocabulary;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class ListHandlerTestSupport {
+public class ListHandlerTestHelper {
 
     static final NamedResource OWNER = NamedResource.create(Generator.generateUri());
 
@@ -20,16 +16,6 @@ public class ListHandlerTestSupport {
     static final String NEXT_NODE_PROPERTY = "http://krizik.felk.cvut.cz/ontologies/2008/6/sequences.owl#hasNext";
     static final String NODE_CONTENT_PROPERTY =
             "http://krizik.felk.cvut.cz/ontologies/2008/6/sequences.owl#hasContents";
-
-    static <T> void verifyListContent(List<Axiom<T>> expected, List<Axiom<T>> actual) {
-        assertEquals(expected.size(), actual.size());
-        // This is more explicit on failure than just containsAll
-        final Iterator<Axiom<T>> itExp = expected.iterator();
-        final Iterator<Axiom<T>> itAct = actual.iterator();
-        while (itExp.hasNext()) {
-            assertEquals(itExp.next(), itAct.next());
-        }
-    }
 
     static List<NamedResource> generateList() {
         return generateList(5);
